chore: bump alloy 0.9 (#350)

+++ crates/alloy/rpc-types/src/receipt.rs
@@ -43,7 +43,7 @@ impl alloy_network_primitives::ReceiptResponse for OpTransactionReceipt {
         self.inner.transaction_index()
     }
 
-    fn gas_used(&self) -> u128 {
+    fn gas_used(&self) -> u64 {
         self.inner.gas_used()
     }
 
@@ -51,7 +51,7 @@ impl alloy_network_primitives::ReceiptResponse for OpTransactionReceipt {
         self.inner.effective_gas_price()
     }
 
-    fn blob_gas_used(&self) -> Option<u128> {
+    fn blob_gas_used(&self) -> Option<u64> {
         self.inner.blob_gas_used()
     }
 
@@ -71,7 +71,7 @@ impl alloy_network_primitives::ReceiptResponse for OpTransactionReceipt {
         self.inner.authorization_list()
     }
 
-    fn cumulative_gas_used(&self) -> u128 {
+    fn cumulative_gas_used(&self) -> u64 {
         self.inner.cumulative_gas_used()
     }
